Suppress invalid region warning with certain buckets
The AWS SDK attempts to use host-style endpoints when the bucket name
contains only DNS-compliant characters and no periods. (Names with
periods break the SSL wildcard certificate.) Unfortunately, due to the
implementation details of AWS Signature V4, the GetBucketLocation
request now requires path-style endpoints, even when the bucket name is
otherwise compatible with host-style endpoints. The AWS SDK has a bug,
aws/aws-sdk-ruby#1374, that causes GetBucketLocation requests to be made
with host-style endpoints instead of path-style endpoints. The AWS SDK
catches the error and retries with the appropriate request, but prints
an annoying warning in the process.

This commit works around the issue for now, thus fixing #31.